code review: refactor out `GetInboundRouteName()` function
